backend/requirements.txt
fastapi
uvicorn
openai
deep-translator
python-multipart
python-dotenv
gunicorn
motor
pytest
pytest-asyncio
httpx

